feat: move jackson bus configuration helpers in a separate artifact, fix #547
