+++ core/taskpool/pom.xml
@@ -38,6 +38,11 @@
       </dependency>
       <dependency>
         <groupId>io.holunda.polyflow</groupId>
+        <artifactId>polyflow-bus-jackson</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>io.holunda.polyflow</groupId>
         <artifactId>polyflow-datapool-api</artifactId>
         <version>${project.version}</version>
       </dependency>
