+++ core/taskpool/taskpool-core/pom.xml
@@ -21,6 +21,10 @@
       <groupId>io.holunda.polyflow</groupId>
       <artifactId>polyflow-taskpool-event</artifactId>
     </dependency>
+    <dependency>
+      <groupId>io.holunda.polyflow</groupId>
+      <artifactId>polyflow-bus-jackson</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>org.springframework.boot</groupId>
