+++ integration/common/taskpool-sender/pom.xml
@@ -21,7 +21,10 @@
       <groupId>io.holunda.polyflow</groupId>
       <artifactId>polyflow-variable-serializer</artifactId>
     </dependency>
-
+    <dependency>
+      <groupId>io.holunda.polyflow</groupId>
+      <artifactId>polyflow-bus-jackson</artifactId>
+    </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter</artifactId>
